TC-7: Refactored code: Modularized configuration handling and utils. PyLint score increased to 9+ #TC-7
